+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Made space between statistics buttons and visualization options more.
 - Updated the functionality of CBC Solver to make it more effective in comprehending all the constraints.
 - Renewed the solution timers to be more user friendly by including both milliseconds and the combination of minutes and seconds.
+- Updated the table results display for better details of statistics comparison between multiple solvers' solutions.
 
 ### Refactoring
 - Updated code structure to be more clean through Separation of Concerns.